Resolved issue #482 - Player stats:
Although commit 663c711 fixed an important issue, it also caused one.
Aura modifiers were applied even though they did not stack.
Even though the aura itself did not stack, the modifiers did,
accumulating bonuses infinitely.

This patch, as well as the patch it updates are currently not considered stable as they might
cause more problems that they fix, and should be submitted to extensive testing.

--HG--
branch : phase

+++ src/collision/BIH.h
@@ -175,7 +175,7 @@ class BIH
                 while (true)
                 {
                     // Check for element existence before accessing position
-                    if (tree.size() < 1 || node >= tree.size()) 
+                    if (tree.size() < 1 || uint32(node) >= tree.size()) 
                         break;
 
                     uint32 tn = tree[node];
